+++ entities/Exchanges/BinanceSpotExchange.py
@@ -7,11 +7,15 @@ import ccxt
 import toml
 import ccxt.async_support
 
+from entities.Exchanges.SpotExchange import SpotExchange
 
-class BinanceSpotExchange(BitfinexSpotExchange):
+
+class BinanceSpotExchange(SpotExchange):
+    IS_SUPPORT_HIDDEN: bool = False
+    MAX_ORDER_LIMIT = 50
+
     def __init__(self, api_key, api_secret, name, proxy, nonce_multiplier):
-        super().__init__(api_key, api_secret, name, proxy, nonce_multiplier)
-        # ccxt.bitfinex
+        super().__init__(api_key, api_secret, name)
         self.proxy = proxy
         self.logger = create_logger()
         self.client = ccxt.binance({
@@ -23,28 +27,26 @@ class BinanceSpotExchange(BitfinexSpotExchange):
         self.client.set_sandbox_mode(config['Testnet'])
         self.client.options['warnOnFetchOpenOrdersWithoutSymbol'] = False
 
-        # self.async_client = ccxt.async_support.binance({
-        #     'apiKey': self.api_key,
-        #     'secret': self.api_secret,
-        #     'proxies': self.proxy
-        # })
-        # self.async_client.set_sandbox_mode(config['Testnet'])
-        # self.async_client.options['warnOnFetchOpenOrdersWithoutSymbol'] = False
-
     def create_order(self, symbol, side, price, qty, ord_type='GTC', **kwargs):
-        response = self.client.create_order(symbol=symbol, side=side, price=price, amount=qty, type='limit')
+        _type = kwargs['type'] if 'type' in kwargs else 'limit'
+        params = {'timeInForce': ord_type}
+        if _type == 'market':
+            params={}
+        response = self.client.create_order(symbol=symbol, side=side, price=price, amount=qty, type=_type, params=params)
         return response
 
     def market_order(self, symbol, side, qty):
         response = self.client.create_market_order(symbol=symbol, side=side, amount=qty)
         return response
 
-    def multi_market_orders(self, symbol, side, list_volume, _type="EXCHANGE IOC", **kwargs):
+    def multi_market_orders(self, symbol, side, list_volume, _type='limit', **kwargs):
         result = dict()
         pool = ThreadPool(len(list_volume))
+        i = 0
         for level in list_volume:
-            result[level[0]] = pool.apply_async(self.create_order,
+            result[i] = pool.apply_async(self.create_order,
                                                (symbol, side, level[0], level[1]), kwargs)
+            i+=1
         result = {k: v.get() for k, v in result.items()}
         return result
